file re-org
set in place new |printhelp|
